sligtly less hacky
