+++ tests/plugins/test_plugins.py
@@ -3,33 +3,16 @@ import unittest
 from straxen.test_utils import nt_test_run_id
 import os
 import shutil
-from importlib import import_module
+from _core import PluginTestAccumulator
 
-# Call all tests starting with this name
-_magic_key = 'plugin_test_'
+# Need import to attach new tests to the PluginTestAccumulator
+import event_building
 
 
-# Somehow the registering cases some issues with importing this main
-# class elsewhere (pytest won't recognize the setUpClass anymore, ideas
-# to fix are welcome)
-@unittest.skipIf(not __file__.endswith('test_plugins.py'),
-                 'Only run from main testing file')
 # Don't bother with remote tests
 @unittest.skipIf(not straxen.utilix_is_configured(), "No db access, cannot test!")
-class PluginTest(unittest.TestCase):
+class PluginTest(unittest.TestCase, PluginTestAccumulator):
     exclude_plugins = 'events_sync_mv', 'events_sync_nv'
-
-    @classmethod
-    def register(cls, test_name, func=None):
-        # See URLConfigs for the original insparation.
-        def wrapper(func):
-            if not isinstance(test_name, str):
-                raise ValueError('test_name name must be a string.')
-
-            setattr(cls, test_name, func)
-            return func
-
-        return wrapper(func) if func is not None else wrapper
 
     @classmethod
     def setUpClass(cls) -> None:
@@ -40,7 +23,6 @@ class PluginTest(unittest.TestCase):
         """
         cls.st = straxen.test_utils.nt_test_context()
         cls.run_id = nt_test_run_id
-        # cls.st.make(cls.run_id, 'records')
 
     @classmethod
     def tearDownClass(cls) -> None:
@@ -56,20 +38,10 @@ class PluginTest(unittest.TestCase):
 for _target in straxen.test_utils.nt_test_context()._plugin_class_registry.keys():
     if _target in PluginTest.exclude_plugins:
         continue
-
-    # pylint disable=cell-var-from-loop
-    @PluginTest.register(f'test_{_target}')
-    def _make(self, target=_target):
-        self.st.make(self.run_id, target)
-
-# There is probably a cleaner step but add all the functions in any .py
-# file in this dir to the tests
-for _file in os.listdir(os.path.dirname(os.path.abspath(__file__))):
-    if not _file.endswith('.py') and _file != str(__file__):
+    if _target != 'records':
         continue
 
-    _module = import_module(_file[:-len('.py')])
-
-    for attr, _object in _module.__dict__.items():
-        if attr.startswith(_magic_key):
-            PluginTest.register(f'test_{attr[len(_magic_key):]}', _object)
+    # pylint disable=cell-var-from-loop
+    @PluginTestAccumulator.register(f'test_{_target}')
+    def _make(self, target=_target):
+        self.st.make(self.run_id, target)
